requirements.txt
fastapi
uvicorn
playwright
google-genai
python-dotenv
beautifulsoup4
pydantic
pydantic-settings
loguru
httpx
pytest
pytest-asyncio
